Промежуточная версия Переделаны api: content и newspage

+++ data/API/ContentAPI/ContentResource.py
@@ -7,6 +7,7 @@ from data.content import Content
 from data.API.ContentAPI.parser_content import parser_content
 from data.smartpage import Smartpage
 from data.API.AuditlogAPI.AuditlogResource import add_auditlog
+from main import password_manager
 
 
 def raise_error(error):
@@ -38,47 +39,34 @@ def find_by_id(id, session):
 
 
 class ContentResource(Resource):
-    def get(self, email, password, content_id):
-        admin, session = check_admin_status(email, password)
-        content, session = find_by_id(content_id, session)
-        return jsonify(content.to_dict(
-            only=('id', 'position', 'heading', 'type', 'image', 'animation_type', 'text', 'tags', 'author_id', 'smartpage_id')))
-
-    def delete(self, email, password, content_id):
-        admin, session = check_admin_status(email, password)
-        content, session = find_by_id(content_id, session)
-        heading, c_pos, type = content.heading, 1, content.type
-        session.delete(content)
-        blocks = session.query(Content).filter(Content.smartpage_id == content.smartpage_id).order_by(
-            Content.position).all()
-        for block in blocks:
-            block.position, c_pos = c_pos, c_pos + 1
-        session.commit()
-        add_auditlog("Удаление",
-                     f"{admin.name} {admin.surname} удаляет блок контента {heading}, с типом данных: {type}",
-                     admin, datetime.datetime.now())
-        return jsonify({"success": f"Блок контента {heading} успешно удален"})
-
-    def put(self, email, password, content_id):
-        admin, session = check_admin_status(email, password)
-        content, session = find_by_id(content_id, session)
+    def put(self, content_id):
         args, count = parser_content.parse_args(), 0
-        cont_dict = content.to_dict(only=('id', 'position', 'heading', 'type', 'image', 'animation_type', 'text', 'tags'))
-        keys = list(
-            filter(lambda key: args[key] is not None and args[key] != cont_dict[key] and key in list(cont_dict.keys()),
-                   list(args.keys())))
-        for key in list(args.keys()):
-            if args[key] is not None and args[key] != cont_dict[key]:
+        if not all(args[key] is not None for key in ['admin_email', 'action']):
+            raise_error('Пропущены некоторые важные аргументы')
+        admin, session = check_admin_status(args["admin_email"], password_manager.get_password(args["admin_email"]))
+        content, session = find_by_id(content_id, session)
+        if args['action'] == "get":
+            return jsonify(content.to_dict(only=('id', 'position', 'heading', 'type', 'image', 'animation_type', 'text', 'tags', 'author_id', 'smartpage_id')))
+        elif args["action"] == "delete":
+            c_pos = 1
+            session.delete(content)
+            blocks = session.query(Content).filter(Content.smartpage_id == content.smartpage_id).order_by(
+                Content.position).all()
+            for block in blocks:
+                block.position, c_pos = c_pos, c_pos + 1
+            session.commit()
+            add_auditlog("Удаление", f"{admin.name} {admin.surname} удаляет блок контента {content.heading}, "
+                                     f"с типом данных: {content.type}", admin, datetime.datetime.now())
+            return jsonify({"success": f"Блок контента {content.heading} успешно удален"})
+        elif args["action"] == "put":
+            cont_dict = content.to_dict(only=('position', 'heading', 'type', 'image', 'animation_type', 'text', 'tags'))
+            keys = list(filter(lambda key: args[key] is not None and key in list(cont_dict.keys()) and args[key] != cont_dict[key], list(args.keys())))
+            for key in keys:
                 count += 1
-                if key == 'id':
-                    if session.query(Content).filter(Content.id == args["id"]).first():
-                        raise_error("Этот id уже занят")
-                    content.id = args['id']
                 if key == "position":
                     if args["position"] < 1:
                         args["position"] = 1
-                    max_pos = session.query(Content).filter(Content.smartpage_id == content.smartpage_id).order_by(
-                        Content.position).all()[-1].position
+                    max_pos = session.query(Content).filter(Content.smartpage_id == content.smartpage_id).order_by(Content.position).all()[-1].position
                     if args["position"] > max_pos:
                         pos = 1
                         content.position = max_pos + 1
@@ -104,17 +92,17 @@ class ContentResource(Resource):
                     content.tags = args["tags"]
                 if key == "heading":
                     content.heading = args["heading"]
-        if count == 0:
-            return raise_error("Пустой запрос")
-        cont_dict_2 = content.to_dict(only=('id', 'position', 'heading', 'type', 'image', 'animation_type', 'text', 'tags'))
-        list_chang = [
-            f'изменяет {key} с {cont_dict[key]} на {cont_dict_2[key]}' if key != "image" else "изменяет изображения" for
-            key in keys]
-        # print([block.type for block in session.query(Content).order_by(Content.position).all()])
-        session.commit()
-        add_auditlog("Изменение", f"{admin.name} {admin.surname} изменяет блок контента: {', '.join(list_chang)}",
-                     admin, datetime.datetime.now())
-        return jsonify({"success": f"Блок контента на позиции {content.position} успешно изменен"})
+            if count == 0:
+                return raise_error("Пустой запрос")
+            cont_dict_2 = content.to_dict(only=('position', 'heading', 'type', 'image', 'animation_type', 'text', 'tags'))
+            list_chang = [
+                f'изменяет {key} с {cont_dict[key]} на {cont_dict_2[key]}' if key != "image" else "изменяет изображения" for
+                key in keys]
+            session.commit()
+            add_auditlog("Изменение", f"{admin.name} {admin.surname} изменяет блок контента: {', '.join(list_chang)}",
+                         admin, datetime.datetime.now())
+            return jsonify({"success": f"Блок контента на позиции {content.position} успешно изменен"})
+        raise_error("Неизвестный метод")
 
 
 class ContentListRecourse(Resource):
@@ -136,11 +124,11 @@ class ContentListRecourseId(Resource):
 
 
 class CreateContentResource(Resource):
-    def post(self, email, password):
-        admin, session = check_admin_status(email, password)
+    def post(self):
         args = parser_content.parse_args()
-        if not all(args[key] is not None for key in ['type', 'page_id', 'heading']):
+        if not all(args[key] is not None for key in ['type', 'page_id', 'heading', 'admin_email']):
             raise_error('Пропущены некоторые аргументы, необходимые для создания страницы')
+        admin, session = check_admin_status(args['admin_email'], password_manager.get_password(args['admin_email']))
         page = session.query(Smartpage).get(args["page_id"])
         if page is None:
             raise_error(f"Страница с id {args['page_id']} не найдена")
